# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Jogos",
      "version": "1.0.0",
      "description": "API para cadastro de jogos"
    },
    "paths": {
      "/jogos": {
        "get": {
          "summary": "Obter a lista de jogos",
          "responses": {
            "200": {
              "description": "Lista de jogos retornada com sucesso"
            }
          }
        },
        "post": {
          "summary": "Criar um novo jogo",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jogos"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description":  "Novo jogo cadastrada com sucesso!"
            },
            "400": {
              "description": "ID já existe, tente novamente com um ID diferente"
            }
          }
        }
      },
      "/jogos/{id}": {
        "put": {
          "summary": "Atualizar informações de um jogo pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jogos"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Informações do jogo atualizadas com sucesso"
            },
            "404": {
              "description": "jogo não encontrado"
            }
          }
        },
        "delete": {
          "summary": "Excluir um jogo pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Jogo excluído com sucesso"
            },
            "404": {
              "description": "Jogo não encontrado"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "jogos": {
          "type": "object",
          "properties": {
            "id": {
                "type": "integer"
              },
            "nome": {
              "type": "string"
            },
            "dataLancamento": {
              "type": "string"
            },
            "descricao": {
                "type": "string"
              },
              "imagem": {
                "type": "string"
              }
          }
        }
      }
    }
  }